transport2030: move GPlanning to common
