+++ src/climatevision/generator/common/g.py
@@ -100,3 +100,34 @@ class GConsult(G):
             invest_pa_com=invest_pa_com,
             ratio_wage_to_emplo=ratio_wage_to_emplo,
         )
+
+
+@dataclass(kw_only=True)
+class GPlanning(G):
+    ratio_wage_to_emplo: float
+
+    @classmethod
+    def calc(cls, inputs: Inputs, invest: float) -> "GPlanning":
+        ass = inputs.ass
+        entries = inputs.entries
+
+        ratio_wage_to_emplo = ass("Ass_T_C_yearly_costs_per_planer")
+        invest_com = invest * ass("Ass_T_C_ratio_public_sector_100")
+        invest_pa_com = invest_com / entries.m_duration_target
+        invest_pa = invest / entries.m_duration_target
+        cost_wage = invest_pa
+        demand_emplo = div(cost_wage, ratio_wage_to_emplo)
+        demand_emplo_new = demand_emplo
+        demand_emplo_com = demand_emplo_new
+
+        return cls(
+            cost_wage=cost_wage,
+            demand_emplo=demand_emplo,
+            demand_emplo_com=demand_emplo_com,
+            demand_emplo_new=demand_emplo_new,
+            invest=invest,
+            invest_com=invest_com,
+            invest_pa=invest_pa,
+            invest_pa_com=invest_pa_com,
+            ratio_wage_to_emplo=ratio_wage_to_emplo,
+        )
